requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
python-dotenv